Added one more test for the player and some new mock responses.
CAUTION: the test added has been commented because it depends on your library. You have to customize it bebore you run it.

CAUTION: For an unknown reason, the groupsActivity is started before the mainActivity. This might be the cause of tests not passing when you run them all together. Try running them separately.